+++ vertex.h
@@ -1,4 +1,4 @@
-//1761414855B69983BD8035097EFBD312EB0527F0
+// 1761414855B69983BD8035097EFBD312EB0527F0
 
 #ifndef VERTEX_H
 #define VERTEX_H
@@ -6,19 +6,21 @@
 #include <iostream>
 #include <cmath>
 
-struct Vertex {
+struct Vertex
+{
     int x, y;
     std::size_t i;
     int area();
-    bool operator== (Vertex const &other) const;
-    bool operator!= (Vertex const &other) const;
-    bool operator< (Vertex const &other) const;
-    int pow_dist (Vertex const &other) const;
-    int insert_dist (Vertex const &before, Vertex const &after) const;
+    bool operator==(Vertex const &other) const;
+    bool operator!=(Vertex const &other) const;
+    bool operator<(Vertex const &other) const;
+    uint64_t pow_dist(Vertex const &other) const;
+    uint64_t insert_dist(Vertex const &before, Vertex const &after) const;
 };
 
-struct VertexComp {
-    bool operator() (const Vertex &a, const Vertex &b) const;
+struct VertexComp
+{
+    bool operator()(const Vertex &a, const Vertex &b) const;
 };
 
 #endif
